refactor: directive with store
