+++ apps/di/src/app/app.component.ts
@@ -2,9 +2,8 @@
 import { AsyncPipe, NgFor } from '@angular/common';
 import { Component, Directive } from '@angular/core';
 import { TableModule } from 'primeng/table';
-import { CurrencyPipe } from './currency.pipe';
-import { CurrencyService } from './currency.service';
 import { Product, products } from './product.model';
+import { CurrencyCodeDirective } from './currency-code.directive';
 
 interface ProductContext {
   $implicit: Product;
@@ -25,8 +24,13 @@ export class ProductDirective {
 
 @Component({
   standalone: true,
-  imports: [TableModule, CurrencyPipe, AsyncPipe, NgFor, ProductDirective],
-  providers: [CurrencyService],
+  imports: [
+    TableModule,
+    CurrencyCodeDirective,
+    AsyncPipe,
+    NgFor,
+    ProductDirective,
+  ],
   selector: 'app-root',
   template: `
     <p-table [value]="products">
@@ -40,9 +44,9 @@ export class ProductDirective {
       <ng-template pTemplate="body" let-product>
         <tr>
           <td>{{ product.name }}</td>
-          <td>{{ product.priceA | currency | async }}</td>
-          <td>{{ product.priceB | currency | async }}</td>
-          <td>{{ product.priceC | currency | async }}</td>
+          <td [code]="product.currencyCode">{{ product.priceA }}</td>
+          <td [code]="product.currencyCode">{{ product.priceB }}</td>
+          <td [code]="product.currencyCode">{{ product.priceC }}</td>
         </tr>
       </ng-template>
     </p-table>
